Added weather update and addition logic

+++ app/src/main/java/com/example/danieljackson/weatherapp/data/network/model/Wind.java
@@ -10,10 +10,10 @@ public class Wind {
     private Double speedKilometersPerHour;
 
     @SerializedName("deg")
-    private Integer degree;
+    private Double degree;
 
     @VisibleForTesting
-    public Wind(Double speedKilometersPerHour, Integer degree) {
+    public Wind(Double speedKilometersPerHour, Double degree) {
         this.speedKilometersPerHour = speedKilometersPerHour;
         this.degree = degree;
     }
@@ -23,6 +23,6 @@ public class Wind {
     }
 
     public Integer getDegree() {
-        return degree;
+        return degree.intValue();
     }
 }
